Refactor: Move dummy data to json file
